AE - Add message type column to Specification entity and dropdown
A message type column has been added to the specification entity, in order to allow for future filtering by message type, and prevent erroneous specification naming. A dropdown with message types has been added to the specification screen, which is part of the aforementioned prevention of erroneous naming of new specifications.