AE - Change specification CSV functions to run in service class
The function to rebuild the CSV of a specification has been moved to the SpecificationRowService class, allowing for its reuse in other controller methods soon to be implemented. The error code setter has also been adapted so that rows are found directly within the service, rather than in the controller, and passed in as a method.